backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
opencv-python>=4.8.0
mediapipe>=0.10.0  # For face landmarks
torch>=2.0.0
transformers>=4.30.0
pydantic>=2.0.0
starlette>=0.27.0
